+++ imports/channel/channel-users.js
@@ -4,15 +4,6 @@ ChannelUsers = new Mongo.Collection('channel-users');
 
 Channels = Meteor.users;
 
-Channels.helpers({
-  className() {
-    if (this.profile && this.profile.channel) {
-      return 'Channels';
-    }
-    return 'Users';
-  }
-});
-
 ChannelUsers.helpers({
   user() {
     return Users.findOne(this.userId);