+++ server/main.js
@@ -1,13 +1,15 @@
+// basic
 import '/imports/model/server';
 import '/imports/i18n/server';
+import '/imports/startup/server';
 import '/imports/files/server';
 import '/imports/chat/server';
 import '/imports/email/server';
 import '/imports/website/server';
 import '/imports/mailgun/server';
-import '/imports/startup/server';
+import '/imports/profile/server';
+// modules
 import '/imports/channel/server';
 import '/imports/roster/server';
 import '/imports/charts/server';
-import '/imports/profile/server';
 import '/imports/cordova/server';